fix: refactored MP3 file tests
