--- texttospeech/ssml_addresses/README.rst
@@ -1,97 +1,0 @@
-.. This file is automatically generated. Do not edit this file directly.
-
-Google Cloud Text-to-Speech API 'Speaking Addresses with SSML' Tutorial Python Samples
-===============================================================================
-
-.. image:: https://gstatic.com/cloudssh/images/open-btn.png
-   :target: https://console.cloud.google.com/cloudshell/open?git_repo=https://github.com/GoogleCloudPlatform/python-docs-samples&page=editor&open_in_editor=/README.rst
-
-
-This directory contains samples for Google Cloud Text-to-Speech API 'Speaking Addresses with SSML' Tutorial. Google Cloud Text-To-Speech API 'Speaking Addresses with SSML Tutorial'_ enables you to use Speech Synthesis Markup Language (SSML) to speak a text file of addresses. You can embed SSML commands in a string of text to personalize synthetic audio from Cloud Text-to-Speech API.
-
-
-
-
-.. _Google Cloud Text-to-Speech API 'Speaking Addresses with SSML' Tutorial: https://cloud.google.com/text-to-speech/docs/ssml-tutorial
-
-Setup
--------------------------------------------------------------------------------
-
-
-Authentication
-++++++++++++++
-
-This sample requires you to have authentication setup. Refer to the
-`Authentication Getting Started Guide`_ for instructions on setting up
-credentials for applications.
-
-.. _Authentication Getting Started Guide:
-    https://cloud.google.com/docs/authentication/getting-started
-
-Install Dependencies
-++++++++++++++++++++
-
-#. Clone python-docs-samples and change directory to the sample directory you want to use.
-
-    .. code-block:: bash
-
-        $ git clone https://github.com/GoogleCloudPlatform/python-docs-samples.git
-
-#. Install `pip`_ and `virtualenv`_ if you do not already have them. You may want to refer to the `Python Development Environment Setup Guide`_ for Google Cloud Platform for instructions.
-
-   .. _Python Development Environment Setup Guide:
-       https://cloud.google.com/python/setup
-
-#. Create a virtualenv. Samples are compatible with Python 2.7 and 3.4+.
-
-    .. code-block:: bash
-
-        $ virtualenv env
-        $ source env/bin/activate
-
-#. Install the dependencies needed to run the samples.
-
-    .. code-block:: bash
-
-        $ pip install -r requirements.txt
-
-.. _pip: https://pip.pypa.io/
-.. _virtualenv: https://virtualenv.pypa.io/
-
-Samples
--------------------------------------------------------------------------------
-
-Speaking addresses with SSML Tutorial
-+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
-.. image:: https://gstatic.com/cloudssh/images/open-btn.png
-   :target: https://console.cloud.google.com/cloudshell/open?git_repo=https://github.com/GoogleCloudPlatform/python-docs-samples&page=editor&open_in_editor=/tts.py,/README.rst
-
-
-
-
-To run this sample:
-
-.. code-block:: bash
-
-    $ python tts.py
-
-
-
-
-The client library
--------------------------------------------------------------------------------
-
-This sample uses the `Google Cloud Client Library for Python`_.
-You can read the documentation for more details on API usage and use GitHub
-to `browse the source`_ and  `report issues`_.
-
-.. _Google Cloud Client Library for Python:
-    https://googlecloudplatform.github.io/google-cloud-python/
-.. _browse the source:
-    https://github.com/GoogleCloudPlatform/google-cloud-python
-.. _report issues:
-    https://github.com/GoogleCloudPlatform/google-cloud-python/issues
-
-
-.. _Google Cloud SDK: https://cloud.google.com/sdk/
