--- texttospeech/ssml_addresses/README.rst.in
@@ -1,19 +1,0 @@
-
-# This file is used to generate README.rst
-
-product:
-  name: Google Cloud Text-to-Speech API 'Speaking Addresses with SSML' Tutorial
-  short_name: Cloud TTS API SSML Addresses Tutorial
-  url: https://cloud.google.com/text-to-speech/docs/ssml-tutorial
-  description: >
-    Google Cloud Text-To-Speech API 'Speaking Addresses with SSML Tutorial'_ enables you to use Speech Synthesis Markup Language (SSML) to speak a text file of addresses. You can embed SSML commands in a string of text to personalize synthetic audio from Cloud Text-to-Speech API.
-
-setup:
-- auth
-- install_deps
-
-samples:
-- name: Speaking addresses with SSML Tutorial
-  file: tts.py
-
-cloud_client_library: true
