--- texttospeech/ssml_addresses/requirements.txt
@@ -1,1 +1,0 @@
-google-cloud-texttospeech==0.4.0
