--- texttospeech/ssml_addresses/resources/example.ssml
@@ -1,4 +1,0 @@
-<speak>123 Street Ln, Small Town, IL 12345 USA
-<break time="2s"/>1 Jenny St &amp; Number St, Tutone City, CA 86753
-<break time="2s"/>1 Piazza del Fibonacci, 12358 Pisa, Italy
-<break time="2s"/></speak>
