--- texttospeech/ssml_addresses/resources/example.txt
@@ -1,3 +1,0 @@
-123 Street Ln, Small Town, IL 12345 USA
-1 Jenny St & Number St, Tutone City, CA 86753
-1 Piazza del Fibonacci, 12358 Pisa, Italy
